# docs/openapi/gateway-v3.yaml
openapi: 3.0.0
info:
  title: Headhunter API Gateway
  version: 1.0.0
  description: 'API Gateway surface aggregating all recruiter platform services.

    OAuth2 access tokens are minted by the Ella identity provider using the

    client-credentials grant and validated by downstream services.

    '
servers:
- url: https://hh-gateway-${ENVIRONMENT}-${REGION}.gateway.dev/
components:
  securitySchemes:
    OAuth2ClientCredentials:
      type: oauth2
      description: OAuth2 client-credentials flow issued by the Ella identity provider.
      flows:
        clientCredentials:
          tokenUrl: https://idp.${ENVIRONMENT}.ella.jobs/oauth/token
          scopes: {}
    TenantApiKey:
      type: apiKey
      name: X-API-Key
      in: header
security:
- OAuth2ClientCredentials: []
  TenantApiKey: []
...
paths:
...
  /v1/admin/refresh-postings:
    post:
      summary: Trigger postings refresh
      description: Publish a postings refresh job and optional scheduler updates.
      operationId: adminRefreshPostings
      tags:
      - Admin
      parameters:
      - $ref: ./schemas/common.yaml#/parameters/TenantIdHeader
      responses:
        '200':
          description: Refresh request accepted with job metadata.
          content:
            application/json:
              schema:
                $ref: ./schemas/common.yaml#/definitions/RefreshResponse
        '202':
          description: Refresh request queued for async execution.
          content:
            application/json:
              schema:
                $ref: ./schemas/common.yaml#/definitions/RefreshResponse
        '429':
          description: Tenant request rate limit exceeded. Retry after the provided
            duration.
          headers:
            Retry-After:
              description: Seconds until the tenant quota window resets.
              type: string
          content:
            application/json:
              schema:
                $ref: ./schemas/common.yaml#/definitions/ErrorResponse
        default:
          $ref: ./schemas/common.yaml#/responses/ErrorResponse
      x-google-backend:
        address: https://${REGION}-run.googleapis.com/apis/serving.knative.dev/v1/namespaces/${PROJECT_ID}/services/hh-admin-svc-${ENVIRONMENT}
        jwt_audience: https://${REGION}-run.googleapis.com/apis/serving.knative.dev/v1/namespaces/${PROJECT_ID}/services/hh-admin-svc-${ENVIRONMENT}
        path_translation: APPEND_PATH_TO_ADDRESS
      x-google-quota:
        metricCosts:
          tenant_global_rps: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas/common.yaml#/definitions/RefreshPostingsRequest
  /v1/admin/refresh-profiles:
    post:
      summary: Trigger profile refresh
      description: Publish a profile refresh job for downstream processors.
      operationId: adminRefreshProfiles
      tags:
      - Admin
      parameters:
      - $ref: ./schemas/common.yaml#/parameters/TenantIdHeader
      responses:
        '200':
          description: Refresh request accepted with job metadata.
          content:
            application/json:
              schema:
                $ref: ./schemas/common.yaml#/definitions/RefreshResponse
        '202':
          description: Refresh request queued for async execution.
          content:
            application/json:
              schema:
                $ref: ./schemas/common.yaml#/definitions/RefreshResponse
        '429':
          description: Tenant request rate limit exceeded. Retry after the provided
            duration.
          headers:
            Retry-After:
              description: Seconds until the tenant quota window resets.
              type: string
          content:
            application/json:
              schema:
                $ref: ./schemas/common.yaml#/definitions/ErrorResponse
        default:
          $ref: ./schemas/common.yaml#/responses/ErrorResponse
      x-google-backend:
        address: https://${REGION}-run.googleapis.com/apis/serving.knative.dev/v1/namespaces/${PROJECT_ID}/services/hh-admin-svc-${ENVIRONMENT}
        jwt_audience: https://${REGION}-run.googleapis.com/apis/serving.knative.dev/v1/namespaces/${PROJECT_ID}/services/hh-admin-svc-${ENVIRONMENT}
        path_translation: APPEND_PATH_TO_ADDRESS
      x-google-quota:
        metricCosts:
          tenant_global_rps: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./schemas/common.yaml#/definitions/RefreshProfilesRequest
  /v1/admin/snapshots:
    get:
      summary: Data freshness snapshots
      description: Retrieve postings and profiles freshness metrics.
      operationId: adminSnapshots
      tags:
      - Admin
      parameters:
      - $ref: ./schemas/common.yaml#/parameters/TenantIdHeader
      - name: tenantId
        in: query
        required: false
        type: string
      responses:
        '200':
          description: Snapshots response with freshness data.
          content:
            application/json:
              schema:
                $ref: ./schemas/common.yaml#/definitions/SnapshotsResponse
        '429':
          description: Tenant request rate limit exceeded. Retry after the provided
            duration.
          headers:
            Retry-After:
              description: Seconds until the tenant quota window resets.
              type: string
          content:
            application/json:
              schema:
                $ref: ./schemas/common.yaml#/definitions/ErrorResponse
        default:
          $ref: ./schemas/common.yaml#/responses/ErrorResponse
      x-google-backend:
        address: https://${REGION}-run.googleapis.com/apis/serving.knative.dev/v1/namespaces/${PROJECT_ID}/services/hh-admin-svc-${ENVIRONMENT}
        jwt_audience: https://${REGION}-run.googleapis.com/apis/serving.knative.dev/v1/namespaces/${PROJECT_ID}/services/hh-admin-svc-${ENVIRONMENT}
        path_translation: APPEND_PATH_TO_ADDRESS
      x-google-quota:
        metricCosts:
          tenant_global_rps: 1
...